+++ transport/tchannel/options.go
@@ -22,12 +22,15 @@ package tchannel
 
 import (
 	"context"
+	"crypto/tls"
 	"net"
 	"time"
 
 	"github.com/opentracing/opentracing-go"
 	"github.com/uber/tchannel-go"
+	"go.uber.org/net/metrics"
 	backoffapi "go.uber.org/yarpc/api/backoff"
+	yarpctls "go.uber.org/yarpc/api/transport/tls"
 	"go.uber.org/yarpc/internal/backoff"
 	"go.uber.org/zap"
 )
@@ -53,6 +56,7 @@ type transportOptions struct {
 	ch                             Channel
 	tracer                         opentracing.Tracer
 	logger                         *zap.Logger
+	meter                          *metrics.Scope
 	addr                           string
 	listener                       net.Listener
 	dialer                         func(ctx context.Context, network, hostPort string) (net.Conn, error)
@@ -62,6 +66,8 @@ type transportOptions struct {
 	originalHeaders                bool
 	nativeTChannelMethods          NativeTChannelMethods
 	excludeServiceHeaderInResponse bool
+	inboundTLSConfig               *tls.Config
+	inboundTLSMode                 *yarpctls.Mode
 }
 
 // newTransportOptions constructs the default transport options struct
@@ -94,6 +100,15 @@ func Tracer(tracer opentracing.Tracer) TransportOption {
 func Logger(logger *zap.Logger) TransportOption {
 	return func(t *transportOptions) {
 		t.logger = logger
+	}
+}
+
+// Meter sets a meter to use for transport metrics.
+//
+// The default is to not emit any transport metrics.
+func Meter(meter *metrics.Scope) TransportOption {
+	return func(t *transportOptions) {
+		t.meter = meter
 	}
 }
 
@@ -226,5 +241,22 @@ func WithNativeTChannelMethods(nativeMethods NativeTChannelMethods) TransportOpt
 func ExcludeServiceHeaderInResponse() TransportOption {
 	return func(option *transportOptions) {
 		option.excludeServiceHeaderInResponse = true
+	}
+}
+
+// InboundTLSMode return TransportOption that sets inbound TLS mode.
+// It must be noted that TLS configuration must be passed separately using
+// option InboundTLSConfiguration.
+func InboundTLSMode(mode yarpctls.Mode) TransportOption {
+	return func(option *transportOptions) {
+		option.inboundTLSMode = &mode
+	}
+}
+
+// InboundTLSConfiguration returns TransportOption that provides the TLS
+// configuration used for setting up TLS inbound.
+func InboundTLSConfiguration(tlsConfig *tls.Config) TransportOption {
+	return func(option *transportOptions) {
+		option.inboundTLSConfig = tlsConfig
 	}
 }